requirements.txt
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
attrs==25.3.0
backoff==2.2.1
certifi==2025.7.14
charset-normalizer==3.4.2
contourpy==1.3.2
cycler==0.12.1
Deprecated==1.2.18
distro==1.9.0
fonttools==4.59.0
gql==3.5.3
graphql-core==3.2.6
graphviz==0.21
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
ifcopenshell==0.8.3.post1
isodate==0.7.2
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.8
langchain==0.3.26
langchain-core==0.3.71
langchain-openai==0.3.28
langchain-text-splitters==0.3.8
langgraph==0.5.4
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.5.2
langgraph-sdk==0.1.74
langsmith==0.4.8
lark==1.2.2
matplotlib==3.10.3
multidict==6.6.3
narwhals==1.48.1
networkx==3.5
numpy==2.3.1
openai==1.97.1
orjson==3.11.0
ormsgpack==1.10.0
packaging==25.0
pandas==2.3.1
pillow==11.3.0
plotly==6.2.0
propcache==0.3.2
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
scipy==1.16.0
shapely==2.1.1
six==1.17.0
sniffio==1.3.1
specklepy==3.0.3
SQLAlchemy==2.0.41
tenacity==9.1.2
tiktoken==0.9.0
topologic_core==8.0.1
topologicpy==0.8.41
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
ujson==5.10.0
urllib3==2.5.0
webcolors==24.11.1
websockets==11.0.3
wrapt==1.17.2
xxhash==3.5.0
yarl==1.20.1
zstandard==0.23.0
